requirements.txt
Django
pytest
pytest-playwright
